Organized project and made a proper readme.md document
