+++ composer.json
@@ -15,7 +15,7 @@
         "psr-4": {
             "App\\Core\\": "resources/app/core",
             "App\\Core\\Views\\": "resources/app/core/views",
-            "App\\Core\\Session\\": "resources/app/core/session",
+            "App\\Core\\Sessions\\": "resources/app/core/sessions",
             "App\\Core\\Validation\\": "resources/app/core/validation",
             "App\\Interfaces\\": "resources/app/interfaces",
             "App\\Interfaces\\Views\\": "resources/app/interfaces/views",
